Refactor processing into subprocesses
Using worker threads had the following issues:

1. When the D garbage collector activates, it wants to stop all
threads so that it can safely scan their stack etc. However, when a
thread is in a slow syscall, the GC will block until the thread exits
this syscall. This caused UI to occasionally freeze up when processing
slow disks.

2. We can exit quickly and abandon background subprocesses without any
slow / onerous cleanup. The subprocesses will hit a SIGPIPE as soon as
they finish their current task, which is as quick of an exit as an
explicit shutdown mechanism could make them quit anyway.