btdu.ui.curses: Initial support for negative coordinates
We would like to be able to point the cursor above the window's top
row, and have some writes get discarded until the cursor reaches the
visible rectangle.

This will allow us to put formatted text in a scrollable buffer
without first having to render the entire text to dynamically
allocated memory.