Use FastAppender.peek() instead of .get()
`peek` is the correct method to use when we want a return value that
will be valid until the next `.clear()` / write.

In practice, this means that FastAppender will deallocate the buffers
it discards due to being too small, because we now tell it to not
worry about keeping the slices returned by get() alive.