btdu.ui.browser: Merge Mode.browser with Mode.marks
The previous model was not correct. For example, the UI showed that
one could press [i] to view in full the summary of all marks, but it
was not actually possible to do so.

Now, distinguish the two by checking if the Browser's currentPath
points to the marked summary node (&marked).