btdu.ui.browser: reorder "Selected" and "Viewing" info panes
When a user is exploring the directory hierarchy btdu has built, it is
likely that the user's eyes will be constantly moving between the
cursor and the info pane representing the entry under cursor.

Thus, assuming that the average list height is less than 3/4 of the
screen, having the relevant info pane at the top of the screen leads
to less eye travel distance which is more ergonomic. (Additionally,
if a user is new to btdu, it is likely they will be reading the UI
in a left-to-right, top-to-down fashion, and it again makes more sense
to have the more relevant info panes closer to the beginning of the eye
scan.)