btdu.ui: Fix TOCTOU when rendering delete progress dialog
We need to render it twice, once to measure the size of the dialog,
and a second time to actually put it (centered) on the screen.

It's important that the two renders use the same data, as otherwise
the text size will not match the dialog size, or it can overflow and
trip an assertion / be drawn out of bounds.

Introduce a method that returns the deleter state, and call it only
once when rendering.