btdu.sample: Fix "Unknown inode root" on I/O errors
If a btrfs filesystem is experiencing I/O errors, sampling it with
btdu may have caused it to crash immediately with an "Unknown inode
root" error. The root cause is bad behavior of D's `require` when
combined with a lambda which may throw: because it creates the AA slot
before executing the callback, and never cleans it up if the callback
throws, the slot is permanently left in place as an uninitialized
value. This caused ulterior getRoot calls to return a root ID which
the main process had never seen.