Retain all sharing groups
Replace the per-BrowserNode seenAs HashMap with a global structure
which retains all unique combinations of paths for retrieved samples,
and how many times we've seen them.

This data can act as the canonical source of truth for all other
aggregated metrics. Here, it is used for:

1. Displaying accurate "Shares data with" metrics even when viewing a
   path which is not representative of the sharing group

2. Adjusting aggregated metrics more accurately after deletion

Despite retaining more data, this change lowers memory usage.