btdu.paths: Adjust mark inversion behavior
Make the "invert marks" action (bound to the * key) toggle the marks
on visible nodes only (rather than the parent node so as to include
not-yet-discovered nodes).

Though somewhat divorced from the conceptual model of btdu, this
better aligns with users' expectations, especially when compared to
traditional file managers / disk usage analyzers. Most importantly,
when this action is used to mark all nodes in a directory, only the
nodes visible at that moment will be considered if the user then
invokes the "delete all marked nodes" action (bound to the D key).

The previous behavior is still mostly available by manually toggling
the mark on the parent node (left, space, up, right).