btdu.paths: Fix logic during migration to aggregated data
Fix the logic for deciding which data source to use for the
circumstance where we are currently migrating to storing aggregated
data; in this case, we have multiple children (but no allocated
aggregateData yet).

Additionally, explicitly document / handle the special case where the
tree is brand new, so there are no children nor sharing groups.