btdu.main: Fix ingestion condition logic while rebuilding
While we already skip adding the subprocesses to the socket set if a
rebuild was running at the start of the loop, user action may initiate
a rebuild when we call browser.update(). If this happens, we must not
read from the subprocess file descriptors even if they were signaled
at the start of the loop.